+++ core/src/main/java/github/nighter/smartspawner/spawner/lootgen/LootResult.java
@@ -4,5 +4,5 @@ import org.bukkit.inventory.ItemStack;
 
 import java.util.List;
 
-public record LootResult(List<ItemStack> items, int experience) {
+public record LootResult(List<ItemStack> items, long experience) {
 }
